1.0.8

+++ pwsh/module/buildModule.ps1
@@ -1,11 +1,14 @@
 Write-Host 'Building one file containing all functions'
 Write-Host ' Cleaning build\functions'
-try {
-    Remove-Item -Path .\pwsh\module\build\AzAPICall\functions\AzAPICallFunctions.ps1 -Verbose -ErrorAction Stop
-}
-catch {
-    Write-Host ' Cleaning build/functions failed'
-    throw
+
+if (Test-Path .\pwsh\module\build\AzAPICall\functions\AzAPICallFunctions.ps1) {
+    try {
+        Remove-Item -Path .\pwsh\module\build\AzAPICall\functions\AzAPICallFunctions.ps1 -Verbose -ErrorAction Stop
+    }
+    catch {
+        Write-Host ' Cleaning build/functions failed'
+        throw
+    }
 }
 
 Get-ChildItem -path .\pwsh\module\dev\AzAPICall\functions | ForEach-Object -Process {
